victini | archiver: impl mstate_read and optimize

+++ src/hzip/core/blob/hzblob.h
@@ -9,7 +9,7 @@
 #include <hzip/utils/common.h>
 
 struct hz_mstate: public hz_mem_iface {
-    uint64_t *data;
+    uint8_t *data;
     uint64_t length;
 
     hz_mstate() {
